Committing an almost working state.
At this point, all the difficult stuff has been figured out.  I want the
app to be highly available, which means rolling deployments, which means
multiple web processes.

A database instance is necessary for data storage to be accessible
between web processes. Currently it's only being used to store the
active games, but I see no need to cleanup this list, as it would be
neat to keep game history anyway.

WebSockets are used to allow real-time communication between client and
server and vice-versa, but there needed to be a way for separate web
processes to talk to allow user A on web1 to talk to user B on web2.

Redis PubSub channels work great for this, and allow users A and B to
communicate without polling. I added an optimization that allows the
PubSub channel to be bypassed if the players are on the same web node.

Not surprisingly, I'm happy with the way the back end code is evolving,
but the front end code seems a mess.