App is in a way more usable state now.
I've separated the tic-tac-toe game logic into a simple module to
prepare for unit testing. I moved the websockets module back into the
game module, because it didn't make sense to have that logic sitting off
on its own.

The communication between client and server is a bit tidier, and
consequently, the client app is a little better organized and getting
better. After every move, a message with the same shape is sent to the
client with all the information needed to represent the game at the
current move. This should make moving to a more reactive ui easier.