Add concept of an "empty piece"
When a cell's content was empty (as in, no piece had been played there)
it sizes differently at different viewport sizes. When a space is empty,
it's now occupied by another letter "-" that's invisible. I don't like
this solution, but at least it didn't require any back end changes.

I'm going to tinker with the css and see if I can find a better
solution.