Need to establish a save point.
This leg of development was dominated by trying to resolve a race
condition that was revealed when trying to match players. Essentially
two players who connect at the same time, will have match records
created with empty player two slots, and if no one else joins, will wait
indefinitely.

In order to solve this, a worker process was created that repeatedly
runs a redis.brpop. So, whenever a player joins, they are picked up by
the worker process, and when enough players are collected, a match is
created and the players notified. This seems to have addressed that
problem, and it doesn't seem like a huge scalability issue because it's
a fast process that only happens once in the lifespan of a match. We'll
see, I guess.

If that wasn't enough, I decided to accommodate other games in the
server logic. They haven't been written yet, and it led me down a bit of
a rabbit hole, but I think it's a more efficient use of resources to
have them on the same server than separate micro servers. Anyway, it was
a fun challenge to solve.

The piece that I haven't yet been able to solve is what happens when a
player disconnects before a match starts. It sets up a condition where a
match is being created with player data that no longer exists. Do I
leave the player data in place and notify the remaining player somehow?
Do I re-queue the existing player to be matched with someone else?