Switch to use ES import/export.
Changing the cjs imports to mjs was straightforward, but there were some
side effects.

Apparently dunder-filename/dirname isn't available in ES modules, so
some extra code had to be used to get a similar variable. I found
examples from the related node pull request, and it works fine, but it's
not exactly an obvious solution.

Some other modules needed to be destructured differently on import,
which, I assume, is because they were still cjs modules.

Since imports are not a sync function (like require), the module that
loads games into a map had to use the async import function, which made
the tests a little more difficult to structure. Other classes had to
change to support the games being passed down to them instead of just
importing them in the module.

The code coverage tool istanbul started outputting an empty report after
the switch. I found some evidence that it may be related to ES modules,
but it has CLI flag that seems to indicate it has support.  Either way,
the same evidence pointed to another tool that's "less hacky" and it
seems to output appropriate information.

I'm not sure why I decided to do this in the middle of these changes,
but I added some metadata to the games. It's a json file that sits
alongside the module files. This seemed appropriate because it's just a
bunch of text, and it doesn't have any bearing on the logic of the
games.