Added a bunch of new stuff.
Built a really efficient algorithm to find the winner for connect four.
Simplified the game state a little. Added winning indicator for connect
four. Decided to ditch redis-mock and just select another db for
testing. The storage in redis is meant to be temporary anyway, but I
just wanted to be safe. The websocket tests aren't working currently,
but I need to save the current state, as it handles client
disconnections well, and I think that change has something to do with
the tests breaking.

I really need to get better about organizing my commits.