Testability improvements
Converted the main entry points to the web application and the lobby to
classes for easier control. This allows the tests easier access to the
internal components.

Decoupled the web sockets server from the pubsub layer for easier
testing and reasoning.

Removed some overly complicated tests, which will soon be replaced with
something simpler.

Added a bunch of new tests.